HDFS-4706. Do not replace root inode for disallowSnapshot.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/branches/HDFS-2802@1469122 13f79535-47bb-0310-9956-ffa450edef68

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/snapshot/SnapshotStats.java
@@ -26,11 +26,11 @@ public interface SnapshotStats {
   /**
    * @return The number of snapshottale directories in the system 
    */
-  public long getNumSnapshottableDirs();
+  public int getNumSnapshottableDirs();
   
   /**
    * @return The number of directories that have been snapshotted
    */
-  public long getNumSnapshots();
+  public int getNumSnapshots();
   
 }
